# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tasks API",
    "version": "1.0.0",
    "description": "API para autenticação de usuários e gerenciamento de tarefas"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registrar usuário",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado"
          },
          "400": {
            "description": "Erro de validação"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login de usuário",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem sucedido"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Listar tarefas",
        "tags": ["Tasks"],
        "responses": {
          "200": {
            "description": "Lista de tarefas"
          }
        }
      },
      "post": {
        "summary": "Criar tarefa",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["pending", "done"] }
                },
                "required": ["title", "description"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Buscar tarefa por ID",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa encontrada"
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      },
      "put": {
        "summary": "Atualizar tarefa",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["pending", "done"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarefa atualizada"
          }
        }
      },
      "delete": {
        "summary": "Deletar tarefa",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Tarefa removida com sucesso"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}